requirements.txt
# CV + GitHub Uygunluk Skoru Projesi
PyGithub>=2.0.0
pdfplumber>=0.10.0
pandas>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0
flask>=3.0.0
streamlit>=1.28.0
